packages/splatkit/pyproject.toml
[project]
name = "splatkit"
dynamic = ["version"]
description = "Core contracts and optional utilities for modular Gaussian splatting."
requires-python = ">=3.11"
dependencies = [
    "beartype>=0.22.9",
    "jaxtyping>=0.3.9",
    "numpy>=2.3.5",
    "plyfile>=1.1.3",
    "pydantic>=2.11.0",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
]

[project.optional-dependencies]
viewer = [
    "marimo-3dv>=0.1.1",
]
data = [
    "huggingface_hub>=0.34.0",
    "opencv-python-headless>=4.10.0",
    "Pillow>=11.0.0",
]
training = []
eval = []
all = [
    "huggingface_hub>=0.34.0",
    "opencv-python-headless>=4.10.0",
    "Pillow>=11.0.0",
]
cu128 = [
    "torch>=2.11.0",
]
cu130 = [
    "torch>=2.11.0",
]

[build-system]
requires = ["hatchling>=1.27.0", "hatch-vcs>=0.5.0"]
build-backend = "hatchling.build"

[tool.uv]
conflicts = [
    [
        { extra = "cu128" },
        { extra = "cu130" },
    ],
]

[tool.uv.sources]
marimo-3dv = { path = "../../third_party/marimo-3dv", editable = true }
torch = [
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-cu130", extra = "cu130" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.hatch.version]
source = "vcs"
fallback-version = "0.1.0a0"

[tool.hatch.version.raw-options]
version_scheme = "no-guess-dev"

[tool.hatch.build.hooks.vcs]
version-file = "src/splatkit/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/splatkit"]
